Statements Without Transactions
Sometimes a statement has no transactions - usually an unused savings account.
As the purpose of this parser is to extract transactions, and describe them by the statement, there is no meaningful way of handling these 'blank' statements.
They're now skipped with an associated message, and there's a log file written to show the skipped (and processed) statements.  The opening and closing balance should be identical in these files, and this is shown against the file on the log.