Operational Project Folders
The folders the user needed for dropping in the pdf statements, or retrieving the exports and logs, had previously been excluded from source control in order to prevent sensitive data ending up on github.  This change excludes the contents of the folders, but add a .gitkeep file to each folder, so the folders themsleves are pushed to github.